Extract last module
